+++ cli/helpers.go
@@ -16,9 +16,7 @@ func tracer() trace.Tracer {
 	return global.Tracer("github.com/yourbase/yb/cli")
 }
 
-func GetTargetPackageNamed(file string) (packages.Package, error) {
-	var targetPackage packages.Package
-
+func GetTargetPackageNamed(file string) (*packages.Package, error) {
 	if file == "" {
 		file = types.MANIFEST_FILE
 	}
@@ -30,29 +28,26 @@ func GetTargetPackageNamed(file string) (packages.Package, error) {
 		_, pkgName := filepath.Split(currentPath)
 		pkg, err := packages.LoadPackage(pkgName, currentPath)
 		if err != nil {
-			return packages.Package{}, fmt.Errorf("Error loading package '%s': %v\n\nSee %s\n", pkgName, err, types.DOCS_URL)
+			return nil, fmt.Errorf("Error loading package '%s': %w\n\nSee %s\n", pkgName, err, types.DOCS_URL)
 		}
-		targetPackage = pkg
-	} else {
-
-		workspace, err := workspace.LoadWorkspace()
-
-		if err != nil {
-
-			return packages.Package{}, fmt.Errorf("Could not find valid configuration: %v\n\nTry running in the package root dir or writing the YML config file (%s) if it is missing. See %s", err, file, types.DOCS_URL)
-		}
-
-		pkg, err := workspace.TargetPackage()
-		if err != nil {
-			return packages.Package{}, fmt.Errorf("Can't load workspace's target package: %v\n\nPackages under this Workspace may be missing a %s file or it's syntax is an invalid YML data. See %s", err, file, types.DOCS_URL)
-		}
-
-		targetPackage = pkg
+		return pkg, nil
 	}
 
-	return targetPackage, nil
+	workspace, err := workspace.LoadWorkspace()
+
+	if err != nil {
+
+		return nil, fmt.Errorf("Could not find valid configuration: %w\n\nTry running in the package root dir or writing the YML config file (%s) if it is missing. See %s", err, file, types.DOCS_URL)
+	}
+
+	pkg, err := workspace.TargetPackage()
+	if err != nil {
+		return nil, fmt.Errorf("Can't load workspace's target package: %w\n\nPackages under this Workspace may be missing a %s file or it's syntax is an invalid YML data. See %s", err, file, types.DOCS_URL)
+	}
+
+	return pkg, nil
 }
 
-func GetTargetPackage() (packages.Package, error) {
+func GetTargetPackage() (*packages.Package, error) {
 	return GetTargetPackageNamed(types.MANIFEST_FILE)
 }